refactor: add utils directory and divide util functions
