+++ src/components/CalendarBodyForMonthView.tsx
@@ -16,9 +16,9 @@ import {
   WeekNum,
 } from '../interfaces'
 import { useTheme } from '../theme/ThemeContext'
-import { typedMemo } from '../utils'
+import { getWeeksWithAdjacentMonths } from '../utils/datetime'
+import { typedMemo } from '../utils/react'
 import { CalendarEventForMonthView } from './CalendarEventForMonthView'
-import { getWeeksWithAdjacentMonths } from '..'
 
 interface CalendarBodyForMonthViewProps<T extends ICalendarEventBase> {
   containerHeight: number
