+++ src/components/CalendarContainer.tsx
@@ -1,6 +1,7 @@
 import dayjs from 'dayjs'
 import React from 'react'
 import { TextStyle, ViewStyle } from 'react-native'
+import { typedMemo } from 'src/utils/react'
 
 import { MIN_HEIGHT } from '../commonStyles'
 import {
@@ -25,8 +26,7 @@ import {
   getDatesInWeek,
   isAllDayEvent,
   modeToNum,
-  typedMemo,
-} from '../utils'
+} from '../utils/datetime'
 import { CalendarBody } from './CalendarBody'
 import { CalendarBodyForMonthView } from './CalendarBodyForMonthView'
 import { CalendarHeader } from './CalendarHeader'
