+++ src/components/CalendarEventForMonthView.tsx
@@ -1,12 +1,13 @@
 import dayjs from 'dayjs'
 import * as React from 'react'
 import { Text, TouchableOpacity, View } from 'react-native'
+import { typedMemo } from 'src/utils/react'
 
 import { u } from '../commonStyles'
 import { useCalendarTouchableOpacityProps } from '../hooks/useCalendarTouchableOpacityProps'
 import { EventCellStyle, EventRenderer, ICalendarEventBase } from '../interfaces'
 import { useTheme } from '../theme/ThemeContext'
-import { getEventSpanningInfo, typedMemo } from '../utils'
+import { getEventSpanningInfo } from '../utils/datetime'
 
 interface CalendarEventProps<T extends ICalendarEventBase> {
   event: T
