+++ src/index.ts
@@ -21,7 +21,9 @@ export * from './interfaces'
 export * from './theme/ThemeContext'
 export * from './theme/ThemeInterface'
 export * from './theme/defaultTheme'
-export * from './utils'
+export * from './utils/datetime'
+export * from './utils/object'
+export * from './utils/react'
 
 export * from './interfaces'
