+++ src/utils/__tests__/utils.test.ts
@@ -4,8 +4,8 @@ import isBetween from 'dayjs/plugin/isBetween'
 import Mockdate from 'mockdate'
 import * as R from 'remeda'
 
-import { ICalendarEventBase } from '../interfaces'
-import * as utils from '../utils'
+import { ICalendarEventBase } from '../../interfaces'
+import * as utils from '../datetime'
 
 Mockdate.set('2021-09-17T04:00:00.000Z')
