+++ src/utils/react.ts
@@ -1,0 +1,3 @@
+import React from 'react'
+
+export const typedMemo: <T>(c: T) => T = React.memo
